Use custom string allocator.
