+++ serialization/serializer.go
@@ -161,7 +161,8 @@ func (s *serializer) flushToDisk(ctx actor.Context) error {
 	}()
 
 	// This maps strings to index position in a slice. This is doing to reduce the file size of the data.
-	strMapToIndex := make(map[string]uint32)
+	// Assume roughly each series has 10 labels.
+	strMapToIndex := make(map[string]uint32, len(s.series)*10)
 	for i, ts := range s.series {
 		ts.FillLabelMapping(strMapToIndex)
 		group.Series[i] = ts
@@ -171,9 +172,9 @@ func (s *serializer) flushToDisk(ctx actor.Context) error {
 		group.Metadata[i] = ts
 	}
 
-	stringsSlice := make([]string, len(strMapToIndex))
+	stringsSlice := make([]types.ByteString, len(strMapToIndex))
 	for stringValue, index := range strMapToIndex {
-		stringsSlice[index] = stringValue
+		stringsSlice[index] = types.ByteString(stringValue)
 	}
 	group.Strings = stringsSlice
