+++ types/serialization_test.go
@@ -28,9 +28,9 @@ func TestLabels(t *testing.T) {
 	strMap := make(map[string]uint32)
 
 	sg.Series[0].FillLabelMapping(strMap)
-	stringsSlice := make([]string, len(strMap))
+	stringsSlice := make([]ByteString, len(strMap))
 	for k, v := range strMap {
-		stringsSlice[v] = k
+		stringsSlice[v] = ByteString([]byte(k))
 	}
 	sg.Strings = stringsSlice
 	buf, err := sg.MarshalMsg(nil)
@@ -45,6 +45,72 @@ func TestLabels(t *testing.T) {
 	for i, lbl := range series2.Labels {
 		require.Equal(t, lbl.Name, series1.Labels[i].Name)
 		require.Equal(t, lbl.Value, series1.Labels[i].Value)
+	}
+}
+
+func BenchmarkFill(b *testing.B) {
+	sg := &SeriesGroup{
+		Series: make([]*TimeSeriesBinary, 0),
+	}
+	for k := 0; k < 1_000; k++ {
+		lblsMap := make(map[string]string)
+		unique := make(map[string]struct{})
+		for j := 0; j < 10; j++ {
+			key := fmt.Sprintf("key_%d", j)
+			v := randString()
+			lblsMap[key] = v
+			unique[key] = struct{}{}
+			unique[v] = struct{}{}
+		}
+		ss := GetTimeSeriesFromPool()
+		ss.Labels = MakeHandles(labels.FromMap(lblsMap))
+
+		sg.Series = append(sg.Series, ss)
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		strMap := make(map[string]uint32, 1_000*10)
+		for _, ts := range sg.Series {
+			ts.FillLabelMapping(strMap)
+		}
+	}
+}
+
+func BenchmarkDeserialize(b *testing.B) {
+	sg := &SeriesGroup{
+		Series: make([]*TimeSeriesBinary, 0),
+	}
+	for k := 0; k < 1_000; k++ {
+		lblsMap := make(map[string]string)
+		for j := 0; j < 10; j++ {
+			key := fmt.Sprintf("key_%d", j)
+			v := randString()
+			lblsMap[key] = v
+		}
+		ss := GetTimeSeriesFromPool()
+		ss.Labels = MakeHandles(labels.FromMap(lblsMap))
+		sg.Series = append(sg.Series, ss)
+	}
+	strMapToIndex := make(map[string]uint32, 1_000*10)
+	for _, ts := range sg.Series {
+		ts.FillLabelMapping(strMapToIndex)
+	}
+	stringsSlice := make([]ByteString, len(strMapToIndex))
+	for stringValue, index := range strMapToIndex {
+		stringsSlice[index] = ByteString(stringValue)
+	}
+	sg.Strings = stringsSlice
+
+	var buf []byte
+	buf, err := sg.MarshalMsg(buf)
+	if err != nil {
+		panic(err)
+	}
+	for i := 0; i < b.N; i++ {
+		_, _, dErr := DeserializeToSeriesGroup(sg, buf)
+		if dErr != nil {
+			panic(dErr.Error())
+		}
 	}
 }
